Move stringifiers to corresponding utility classes
